+++ db/bigtable_common.go
@@ -6,10 +6,10 @@ import (
 	"eth2-exporter/utils"
 	"fmt"
 	"sort"
-	"strings"
 	"time"
 
 	gcp_bigtable "cloud.google.com/go/bigtable"
+	"github.com/sirupsen/logrus"
 )
 
 func (bigtable *Bigtable) WriteBulk(mutations *types.BulkMutations, table *gcp_bigtable.Table, batchSize int) error {
@@ -19,16 +19,22 @@ func (bigtable *Bigtable) WriteBulk(mutations *types.BulkMutations, table *gcp_b
 	ctx, done := context.WithTimeout(context.Background(), time.Minute*5)
 	defer done()
 
-	// pre-sort mutations for efficient bulk inserts
-	sort.Sort(mutations)
-	length := batchSize
 	numMutations := len(mutations.Muts)
 	numKeys := len(mutations.Keys)
-	iterations := numKeys / length
-
 	if numKeys != numMutations {
 		return fmt.Errorf("error expected same number of keys as mutations keys: %v mutations: %v", numKeys, numMutations)
 	}
+
+	// pre-sort mutations for efficient bulk inserts
+	sort.Sort(mutations)
+
+	length := batchSize
+	if length > MAX_BATCH_MUTATIONS {
+		logger.Infof("WriteBulk: capping provided batchSize %v to %v", length, MAX_BATCH_MUTATIONS)
+		length = MAX_BATCH_MUTATIONS
+	}
+
+	iterations := numKeys / length
 
 	for offset := 0; offset < iterations; offset++ {
 		start := offset * length
@@ -38,13 +44,14 @@ func (bigtable *Bigtable) WriteBulk(mutations *types.BulkMutations, table *gcp_b
 		errs, err := table.ApplyBulk(ctx, mutations.Keys[start:end], mutations.Muts[start:end])
 		for _, e := range errs {
 			if e != nil {
-				return err
+				return e
 			}
 		}
-		logger.Infof("%s: wrote from %v to %v rows to bigtable in %.1f s", callingFunctionName, start, end, time.Since(startTime).Seconds())
 		if err != nil {
 			return err
 		}
+		logger.Infof("%s: wrote from %v to %v rows to bigtable in %.1f s", callingFunctionName, start, end, time.Since(startTime).Seconds())
+
 	}
 
 	if (iterations * length) < numKeys {
@@ -60,74 +67,100 @@ func (bigtable *Bigtable) WriteBulk(mutations *types.BulkMutations, table *gcp_b
 			}
 		}
 		logger.Infof("%s: wrote from %v to %v rows to bigtable in %.1fs", callingFunctionName, start, numKeys, time.Since(startTime).Seconds())
-		if err != nil {
-			return err
-		}
+
 		return nil
 	}
 
 	return nil
 }
 
-func (bigtable *Bigtable) DeleteRowsWithPrefix(prefix string) {
+func (bigtable *Bigtable) ClearByPrefix(table string, family, prefix string, dryRun bool) error {
+	if family == "" || prefix == "" {
+		return fmt.Errorf("please provide family [%v] and prefix [%v]", family, prefix)
+	}
 
-	for {
-		ctx, done := context.WithTimeout(context.Background(), time.Second*30)
-		defer done()
+	rowRange := gcp_bigtable.PrefixRange(prefix)
 
-		rr := gcp_bigtable.InfiniteRange(prefix)
+	var btTable *gcp_bigtable.Table
 
-		rowsToDelete := make([]string, 0, 10000)
-		err := bigtable.tableData.ReadRows(ctx, rr, func(r gcp_bigtable.Row) bool {
-			rowsToDelete = append(rowsToDelete, r.Key())
-			return true
-		})
+	switch table {
+	case "data":
+		btTable = bigtable.tableData
+	case "blocks":
+		btTable = bigtable.tableBlocks
+	case "metadata_updates":
+		btTable = bigtable.tableMetadataUpdates
+	case "metadata":
+		btTable = bigtable.tableMetadata
+	case "beaconchain":
+		btTable = bigtable.tableBeaconchain
+	case "machine_metrics":
+		btTable = bigtable.tableMachineMetrics
+	case "beaconchain_validators":
+		btTable = bigtable.tableValidators
+	case "beaconchain_validators_history":
+		btTable = bigtable.tableValidatorsHistory
+	default:
+		return fmt.Errorf("unknown table %v provided", table)
+	}
+
+	mutsDelete := types.NewBulkMutations(MAX_BATCH_MUTATIONS)
+
+	keysCount := 0
+	err := btTable.ReadRows(context.Background(), rowRange, func(row gcp_bigtable.Row) bool {
+
+		if family == "*" {
+			if dryRun {
+				logger.Infof("would delete key %v", row.Key())
+			}
+
+			mutDelete := gcp_bigtable.NewMutation()
+			mutDelete.DeleteRow()
+			mutsDelete.Keys = append(mutsDelete.Keys, row.Key())
+			mutsDelete.Muts = append(mutsDelete.Muts, mutDelete)
+			keysCount++
+		} else {
+			row_ := row[family][0]
+			if dryRun {
+				logger.Infof("would delete key %v", row_.Row)
+			}
+
+			mutDelete := gcp_bigtable.NewMutation()
+			mutDelete.DeleteRow()
+			mutsDelete.Keys = append(mutsDelete.Keys, row_.Row)
+			mutsDelete.Muts = append(mutsDelete.Muts, mutDelete)
+			keysCount++
+		}
+
+		if mutsDelete.Len() == MAX_BATCH_MUTATIONS {
+			logrus.Infof("deleting %v keys (first key %v, last key %v)", len(mutsDelete.Keys), mutsDelete.Keys[0], mutsDelete.Keys[len(mutsDelete.Keys)-1])
+			if !dryRun {
+				err := bigtable.WriteBulk(mutsDelete, btTable, DEFAULT_BATCH_INSERTS)
+
+				if err != nil {
+					logger.Errorf("error writing bulk mutations: %v", err)
+					return false
+				}
+			}
+			mutsDelete = types.NewBulkMutations(MAX_BATCH_MUTATIONS)
+		}
+		return true
+	})
+	if err != nil {
+		return err
+	}
+
+	if !dryRun && mutsDelete.Len() > 0 {
+		logger.Infof("deleting %v keys (first key %v, last key %v)", len(mutsDelete.Keys), mutsDelete.Keys[0], mutsDelete.Keys[len(mutsDelete.Keys)-1])
+
+		err := bigtable.WriteBulk(mutsDelete, btTable, DEFAULT_BATCH_INSERTS)
+
 		if err != nil {
-			logger.WithError(err).WithField("prefix", prefix).Errorf("error reading rows in bigtable_eth1 / DeleteRowsWithPrefix")
-		}
-		mut := gcp_bigtable.NewMutation()
-		mut.DeleteRow()
-
-		muts := make([]*gcp_bigtable.Mutation, 0)
-		for j := 0; j < 10000; j++ {
-			muts = append(muts, mut)
-		}
-
-		l := len(rowsToDelete)
-		if l == 0 {
-			logger.Infof("all done")
-			break
-		}
-		logger.Infof("deleting %v rows", l)
-
-		for i := 0; i < l; i++ {
-			if !strings.HasPrefix(rowsToDelete[i], "1:t:") {
-				logger.Infof("wrong prefix: %v", rowsToDelete[i])
-			}
-			ctx, done := context.WithTimeout(context.Background(), time.Second*30)
-			defer done()
-			if i%10000 == 0 && i != 0 {
-				logger.Infof("deleting rows: %v to %v", i-10000, i)
-				errs, err := bigtable.tableData.ApplyBulk(ctx, rowsToDelete[i-10000:i], muts)
-				if err != nil {
-					logger.WithError(err).Errorf("error deleting row: %v", rowsToDelete[i])
-				}
-				for _, err := range errs {
-					utils.LogError(err, fmt.Errorf("bigtable apply bulk error, deleting rows: %v to %v", i-10000, i), 0)
-				}
-			}
-			if l < 10000 && l > 0 {
-				logger.Infof("deleting remainder")
-				errs, err := bigtable.tableData.ApplyBulk(ctx, rowsToDelete, muts[:len(rowsToDelete)])
-				if err != nil {
-					logger.WithError(err).Errorf("error deleting row: %v", rowsToDelete[i])
-				}
-				for _, err := range errs {
-					utils.LogError(err, "bigtable apply bulk error, deleting remainer", 0)
-				}
-				break
-			}
+			return err
 		}
 	}
 
+	logger.Infof("deleted %v keys", keysCount)
+
+	return nil
 }
