+++ types/eth1.go
@@ -1,0 +1,9 @@
+package types
+
+import "time"
+
+type GetBlockTimings struct {
+	Headers  time.Duration
+	Receipts time.Duration
+	Traces   time.Duration
+}
